Separated the edge into two distinct structs that are linked to each other
The reason for this is so that it's possible to split an edge into two
so we can insert things in the same manner we would a linked list.
Without this, we would need to modify the node's reference to the
input/output edge and we are trying to make each node agnostic about the
content of other nodes.

Now inserting a new node into a current edge is much easier and
eliminating an existing edge is also easier. In the code, the way we act
now is that a the input or output end of an edge is attached to the
exact same node for the entirety of its lifecycle rather than being able
to change where it is pointed to.