Rename the OutputEdge to ReadEdge and InputEdge to WriteEdge
The original names referred to the sides of the edge itself and whether
they accepted an iterator or output an iterator, but that was confusing
because they were always being referred by how they related to the nodes
where the terminology was the exact opposite.

They have been renamed to ReadEdge and WriteEdge so it is more clear
that one side reads from an iterator and the other side writes to an
iterator.